+++ Core/include/backgrounddatabase.h
@@ -7,11 +7,11 @@ class BackgroundDatabase
 {
 public:
   BackgroundDatabase();
-  unsigned long long size();
-  unsigned int get_id(std::string name);
-  std::string get_name_from_id(unsigned int id);
-  BackgroundAsset* get_asset(unsigned int id);
+  size_t size();
+  size_t get_id(std::string name);
+  std::string get_name_from_id(size_t id);
+  BackgroundAsset* get_asset(size_t id);
 private:
   std::vector<BackgroundAsset*> backgrounds;
-  std::map<std::string, unsigned int> id_map;
+  std::map<std::string, size_t> id_map;
 };
