+++ Core/include/tilesetasset.h
@@ -5,12 +5,12 @@
 class TilesetAsset
 {
 public:
-  TilesetAsset(unsigned int id, std::string name, std::string path);
+  TilesetAsset(size_t id, std::string name, std::string path);
   TextureAsset* get_texture();
-  unsigned int get_id();
+  size_t get_id();
   std::string get_name();
 private:
-  unsigned int id;
+  size_t id;
   std::string name;
   TextureAsset texture;
 };
