+++ Core/include/tilesetdatabase.h
@@ -7,12 +7,12 @@ class TilesetDatabase
 {
 public:
   TilesetDatabase();
-  TilesetAsset* get_asset(unsigned int id);
+  TilesetAsset* get_asset(size_t id);
   TilesetAsset* get_asset(std::string name);
-  unsigned int get_id_from_name(std::string name);
-  unsigned long long get_size();
+  size_t get_id_from_name(std::string name);
+  size_t get_size();
 private:
   std::vector<TilesetAsset*> assets;
-  std::map<std::string, unsigned int> id_map;
+  std::map<std::string, size_t> id_map;
 
 };
