+++ Core/source/assets/tilesetasset.cpp
@@ -1,12 +1,12 @@
 #include "tilesetasset.h"
 
-TilesetAsset::TilesetAsset(unsigned int id, std::string name, std::string path) : id(id), name(name), texture(path) {}
+TilesetAsset::TilesetAsset(size_t id, std::string name, std::string path) : id(id), name(name), texture(path) {}
 
 TextureAsset* TilesetAsset::get_texture() {
   return &texture;
 }
 
-unsigned int TilesetAsset::get_id() {
+size_t TilesetAsset::get_id() {
   return id;
 }
