+++ Core/source/databases/backgrounddatabase.cpp
@@ -4,7 +4,7 @@
 BackgroundDatabase::BackgroundDatabase() {
   backgrounds = FileSystem::load_backgrounds();
   for(BackgroundAsset* asset : backgrounds) {
-    id_map.insert(std::pair<std::string, unsigned long long>(asset->get_name(), asset->get_id()));
+    id_map.insert(std::pair<std::string, size_t>(asset->get_name(), asset->get_id()));
   }
 }
 
@@ -12,14 +12,14 @@ unsigned long long BackgroundDatabase::size() {
   return backgrounds.size();
 }
 
-unsigned int BackgroundDatabase::get_id(std::string name) {
+size_t BackgroundDatabase::get_id(std::string name) {
   return id_map.at(name);
 }
 
-std::string BackgroundDatabase::get_name_from_id(unsigned int id) {
+std::string BackgroundDatabase::get_name_from_id(size_t id) {
   return backgrounds.at(id)->get_name();
 }
 
-BackgroundAsset* BackgroundDatabase::get_asset(unsigned int id) {
+BackgroundAsset* BackgroundDatabase::get_asset(size_t id) {
   return backgrounds.at(id);
 }
